Make modern header (see desc)
SCSS has been compiled into a CSS file and used alongside Blade/HTML content to generate a modern header with a hamburger menu and an account access button.

While not yet usable, the buttons will be made functional when the respective sections are created as in line with the Gantt chart.

The hamburger menu will be completed tomorrow as it is the last day to be allocated on the Gantt chart for this topic. The footer will also be completed tomorrow in a matching style.