Add section detail editing
The section details can now be edited using a button at the start of each section. The form will allow the course owner to enter a new title and/or description and submit this data to the server. Existing scripts have been modified such that only the edit or add sections can be open at one time, if the other is open and a button clicked, the other section will automatically be closed.