Add removal/blocking of users from courses
This commit adds the ability in the settings page for the owner to remove users or block them from the course. A new attribute in the user_courses table has been added as a flag for blocking. In order to display the results correctly, the relationships have had primary keys directly specified to prevent errors. The course middleware has also been updated to reflect the blocking flag, which will redirect users with an error message if they are blocked.