Add a facility to add text components to lessons
This will allow a dynamic form to be used, utilising AJAX for live updates to ensure the form is appropriately refreshed without reloading the page. Forms are requested and delivered by the server and using jQuery to present to the client. The server-side processing to add the item to the database has also been implemented.